requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
google-auth==2.39.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
spotipy==2.25.1
urllib3==2.4.0
